ClumsyAlien 0.2 Update merge (#28)
* Updated default to false for breathing, it's only in an Alpha state

* Refactoring code for new Epic's (#10)

* Crafting Table Recipe, Preparing Config for some possible future features

* PR containing 0.1.3 updates and fatal bug patches

* Pre0.1.4, Toggle aim, language updates, model updates / pistol positioning, some sounds but non implemented

* Finalize 0.1.4, pistol muzzle flash, my first endeavor with additional system settings

* Absolutely insane amount of new work

In short

Movement tech, massively improved form CGM experimental, lots of new scope tech, old scopes, recoil adjustment starts (nearly redoing recoil feel effects and attachments), starts of full mag reloading for upcoming Animation support for reloads.

* Large scope tech changes, new recoil systems, fixed xr95 aiming issue, test weapon adjustments (#19)

* Mayday refactoring prep (#21)

* Implemented animations

* Update build.gradle

* Polished ak47 reload animation

* Fix pose translate problem


* Revert "Mayday refactoring prep (#21)"

This reverts commit 64de436c151fab19d267097919631708303af3dd.

* Mayday refactoring prep (#22)

* Implemented animations

* Update build.gradle

* Polished ak47 reload animation

* Fix pose translate problem

* Updated draw animation


* ClumsyAlien animation impl test, qmk glass update

* Mega list of bug fixes, scope model updates, aiming, 2d in gui and much more to be explained later

* Scopes now render 2d mode properly, support for chisel-and-bits mod to replace bullet spots incoming

* Update for future CaB investigation

* Bunch of new weapons animation changes flashlight basics etc, next is to rework weapon recoil duration

* Basic new pistol attachments, new weapons, attachment renders, GUI's and much much more

* Fully Realized pistol attachments, New Rmr optic, Micro Holo rework

* Lots of new weapons, pistol scopes, weapon adjustments, and grenade tech

* redone movement, and tons on tons of recoil systems and such are now added, more explosive tech

* Eotech reticle adjusments to help further increase understanding, some model changes and recoil changs, compacted from failed local changes

* Updated for local new system, updates to reloading, and every single weapon updated for movement speed, most updated for recoil examples

* Quick rebase of movement effects via scientific evidence rather then gameplay

* Smaller optic fix, done for modern 8x and vortex uh1, planned for all optics

* Fixed all reticles under squish conditions, some final model changes, and some temporary

* Added full settings screen accessible bars to adjust onScreenUI, removed flashlight temporarily

* NEAR FINAL CHANGES!

Last upcoming changes

Fixing Server deployment exceptions
VZ61 model update
More sound variants for new weapons

* Refinements, updates before system switch

* RQ update, used to ensure backup between mass weapon balancing

* Final step towards Release? Awaiting final reviews.

* Update for new joining Dev

* Release time baby!

+++ src/main/java/com/tac/guns/entity/ProjectileEntity.java
@@ -4,6 +4,7 @@ package com.tac.guns.entity;
 
 import com.mrcrayfish.obfuscate.common.data.SyncedPlayerData;
 import com.tac.guns.Config;
+import com.tac.guns.GunMod;
 import com.tac.guns.common.BoundingBoxManager;
 import com.tac.guns.common.Gun;
 import com.tac.guns.common.Gun.Projectile;
@@ -25,8 +26,13 @@ import com.tac.guns.util.GunEnchantmentHelper;
 import com.tac.guns.util.GunModifierHelper;
 import com.tac.guns.util.math.ExtendedEntityRayTraceResult;
 import com.tac.guns.world.ProjectileExplosion;
+import mod.chiselsandbits.api.ChiselsAndBitsAPI;
+import mod.chiselsandbits.api.IChiselsAndBitsAPI;
+import mod.chiselsandbits.api.chiseling.ChiselingOperation;
+import mod.chiselsandbits.multistate.mutator.ChiselAdaptingWorldMutator;
 import net.minecraft.block.*;
 import net.minecraft.block.material.Material;
+import net.minecraft.client.Minecraft;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.entity.*;
 import net.minecraft.entity.player.PlayerEntity;
@@ -257,7 +263,7 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
             }
             else
             {
-                this.onHit(result, startVec, endVec);
+                this.onHit(result, startVec, endVec); // Issue
             }
         }
 
@@ -405,6 +411,9 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
 
     private void onHit(RayTraceResult result, Vector3d startVec, Vector3d endVec)
     {
+        if(modifiedGun == null)
+            return;
+
         MinecraftForge.EVENT_BUS.post(new GunProjectileHitEvent(result, this));
 
         if(result instanceof BlockRayTraceResult)
@@ -420,6 +429,9 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
             BlockState state = this.world.getBlockState(pos);
             Block block = state.getBlock();
 
+            if(block.getRegistryName().getPath().contains("_button"))
+                return;
+
             if(Config.COMMON.gameplay.enableGunGriefing.get() && (block instanceof BreakableBlock || block instanceof PaneBlock) && state.getMaterial() == Material.GLASS)
             {
                 this.world.destroyBlock(blockRayTraceResult.getPos(), false);
@@ -430,12 +442,24 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
             //    this.remove();
             //}
 
+/*
             if(block instanceof IDamageable)
             {
-                ((IDamageable) block).onBlockDamaged(this.world, state, pos, this, this.getDamage(), (int) Math.ceil(this.getDamage() / 2.0) + 1);
+                ((IDamageable) block).onBlockDamaged(this.world, state, pos, this, this.getDamage(), (int) Math.ceil(this.getDamage() / 2.0) + 1); // Possibly to help fix button issue and many others, I don't think I can use this for now
+            }
+*/
+
+            if(modifiedGun.getProjectile().isRicochet() &&
+            ((
+                            state.getMaterial() == Material.ROCK ||
+                            state.getMaterial() == Material.PACKED_ICE ||
+                            state.getMaterial() == Material.IRON ||
+                            state.getMaterial() == Material.ANVIL
+            )))
+            {
+                this.onHitBlock(blockRayTraceResult);
             }
 
-            this.onHitBlock(blockRayTraceResult);
             this.onHitBlock(state, pos, blockRayTraceResult.getFace(), hitVec.x, hitVec.y, hitVec.z);
 
             int fireStarterLevel = EnchantmentHelper.getEnchantmentLevel(ModEnchantments.FIRE_STARTER.get(), this.weapon);
@@ -488,6 +512,7 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
         if(headshot)
         {
             damage *= Config.COMMON.gameplay.headShotDamageMultiplier.get();
+            damage *= GunModifierHelper.getAdditionalHeadshotDamage(this.weapon) == 0F ? 1F : GunModifierHelper.getAdditionalHeadshotDamage(this.weapon);
         }
 
         DamageSource source = new DamageSourceProjectile("bullet", this, shooter, weapon).setProjectile();
@@ -507,57 +532,63 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
     {
         if(modifiedGun == null)
             return;
-
-        BlockPos pos = blockRayTraceResult.getPos();
-        BlockState state = this.world.getBlockState(pos);
-
-        if(modifiedGun.getProjectile().isRicochet())
-        {
-            if
-            (
-                    state.getMaterial() == Material.ROCK ||
-                    state.getMaterial() == Material.PACKED_ICE ||
-                    state.getMaterial() == Material.IRON ||
-                    state.getMaterial() == Material.ANVIL
-            )
-            {
-                Direction blockDirection = blockRayTraceResult.getFace();
-                switch (blockDirection) {
-                    case UP:
-                    case DOWN:
-                        this.setMotion(this.getMotion().mul(1, -1, 1));
-                        break;
-                    case EAST:
-                    case WEST:
-                        this.setMotion(this.getMotion().mul(-1, 1, 1));
-                        break;
-                    case NORTH:
-                    case SOUTH:
-                        this.setMotion(this.getMotion().mul(1, 1, -1));
-                        break;
-                    default:
-                        break;
-                }
-
-                Vector3d startVec = this.getPositionVec();
-                Vector3d endVec = startVec.add(this.getMotion());
-
-                EntityResult entityResult = this.findEntityOnPath(startVec, endVec);
-
-                RayTraceResult result = rayTraceBlocks(this.world, new RayTraceContext(startVec, endVec, RayTraceContext.BlockMode.COLLIDER, RayTraceContext.FluidMode.NONE, this), IGNORE_LEAVES);
-
-                if (entityResult != null) {
-                    this.tick();
-                    return;
-                } else {
-                    this.teleportToHitPoint(result);
-                }
-                this.life -= 1;
-            }
+        Direction blockDirection = blockRayTraceResult.getFace();
+        switch (blockDirection) {
+            case UP:
+            case DOWN:
+                this.setMotion(this.getMotion().mul(1, -1, 1));
+                break;
+            case EAST:
+            case WEST:
+                this.setMotion(this.getMotion().mul(-1, 1, 1));
+                break;
+            case NORTH:
+            case SOUTH:
+                this.setMotion(this.getMotion().mul(1, 1, -1));
+                break;
+            default:
+                break;
         }
+
+        Vector3d startVec = this.getPositionVec();
+        Vector3d endVec = startVec.add(this.getMotion());
+        EntityResult entityResult = this.findEntityOnPath(startVec, endVec);
+
+        RayTraceResult result = rayTraceBlocks(this.world, new RayTraceContext(startVec, endVec, RayTraceContext.BlockMode.COLLIDER, RayTraceContext.FluidMode.NONE, this), IGNORE_LEAVES); // Ricochet Raytrace
+
+        if (entityResult != null) {
+            this.tick();
+            return;
+        } else {
+            this.teleportToHitPoint(result);
+        }
+        this.life -= 1;
     }
 
-    protected void onHitBlock(BlockState state, BlockPos pos, Direction face, double x, double y, double z) {
+    private static boolean deleteBitOnHit(BlockPos blockPos, BlockState blockState, double x, double y, double z)//(IParticleData data, double x, double y, double z, Random rand, double velocityMultiplier)
+    {
+        Minecraft mc = Minecraft.getInstance();
+        //IChiselsAndBitsAPI.getInstance()
+        //ChiselAdaptingWorldMutator chiselAdaptingWorldMutator =
+        float bitSize = ChiselsAndBitsAPI.getInstance().getStateEntrySize().getSizePerBit();
+        ChiselsAndBitsAPI.getInstance().getMutatorFactory().in(mc.world, blockPos).overrideInAreaTarget(Blocks.AIR.getDefaultState(), new Vector3d(Math.abs(x),Math.abs(y),Math.abs(z)));
+
+        //ChiselsAndBitsAPI.getInstance().getMutatorFactory().in(mc.world, blockPos).clearInBlockTarget(BlockPos.ZERO, new Vector3d(Math.abs(x),Math.abs(y),Math.abs(z)));//Math.abs(x),Math.abs(y),Math.abs(z) //clearInBlockTarget(BlockPos.ZERO, new Vector3d(Math.abs(x)/bitSize,Math.abs(y)/bitSize,Math.abs(z)/bitSize));
+
+        /*ChiselsAndBitsAPI.getInstance().getMutatorFactory().covering(, blockPos).overrideInAreaTarget(Blocks.AIR.getDefaultState(), new Vector3d(0,0,0));*/
+        //chiselAdaptingWorldMutator.clearInAreaTarget(new Vector3d(Math.abs(x),Math.abs(y),Math.abs(z)));//(BlockPos.ZERO, new Vector3d(Math.abs(x),Math.abs(y),Math.abs(z)));//setInBlockTarget(Blocks.AIR.getDefaultState(), BlockPos.ZERO, new Vector3d(Math.abs(x),Math.abs(y),Math.abs(z))); //clearInBlockTarget(blockPos, new Vector3d(Math.abs(x)/1000,Math.abs(y)/1000,Math.abs(z)/1000));// clearInBlockTarget(blockPos, new Vector3d(x,y,z));
+        return true;
+    }
+
+    protected void onHitBlock(BlockState state, BlockPos pos, Direction face, double x, double y, double z)
+    {
+        /*if(GunMod.cabLoaded)
+        {
+            double holeX = 0.005 * face.getXOffset();
+            double holeY = 0.005 * face.getYOffset();
+            double holeZ = 0.005 * face.getZOffset();
+            deleteBitOnHit(pos, state, holeX, holeY, holeZ);
+        }*/
         PacketHandler.getPlayChannel().send(PacketDistributor.TRACKING_CHUNK.with(() -> this.world.getChunkAt(pos)), new MessageProjectileHitBlock(x, y, z, pos, face));
     }
 
@@ -854,7 +885,7 @@ public class ProjectileEntity extends Entity implements IEntityAdditionalSpawnDa
     }
 
     /**
-     * Author: MrCrayfish
+     * Author: Forked from MrCrayfish, continued by Timeless devs
      */
     public static class EntityResult
     {
